// src/lib/types.ts
export type PlatformId =
  | "twitch"
  | "youtube"
  | "vkplay"
  | "kick"
  | "tiktok"
  | "rutube"
  | "boosty"
  | "donationalerts"
  | "donatepay";